Move FragmentLifecycleEvent into android.events package.
