+++ formula-android/src/main/java/com/instacart/formula/android/internal/FragmentFlowRenderView.kt
@@ -13,7 +13,7 @@ import com.instacart.formula.android.BaseFormulaFragment
 import com.instacart.formula.android.FormulaFragment
 import com.instacart.formula.android.FragmentEnvironment
 import com.instacart.formula.fragment.FragmentFlowState
-import com.instacart.formula.fragment.FragmentLifecycleEvent
+import com.instacart.formula.android.events.FragmentLifecycleEvent
 import com.instacart.formula.android.BackCallback
 import com.instacart.formula.android.FeatureEvent
 import com.instacart.formula.android.FragmentId
